msg actions: Simplify `resolveMultiple` helper function, as `allOf`.
This is simpler to understand, and I think the name is clearer too.

While here, add a type signature.  The most accurate and most natural
type for this is generic in a type parameter; Flow doesn't seem to
have syntax for expressing that in an arrow function, so we depart
from our usual style and write a `function` declaration instead, where
Flow does have syntax for introducing a type parameter.