webview: Add an user-oriented error banner, and switch by default.
For any user who isn't either developing the app or actively giving
beta feedback, there's no value in showing the detailed red error
banner; the information is completely unactionable.  It brings back
memories of those cryptic numbers-filled error messages that desktop
software was notorious for showing in the '90s; not a good look.
This is one recurring thing that users notice and mention to us,
and that causes them to rapidly lose confidence in the app.

On the other hand, there is real value to the user in signaling that
we know an error occurred; it can help them decide to go leave the
screen or kill the app and retry, and help them determine if that
likely worked.

So, add an alternate, "plain" error banner, which is written to be
helpful to the user; and switch to it by default.

What about when a member of the Zulip community sees this message and
wants to help provide details for debugging?  We have a few ways of
getting the information that was in the error object:

 * In a release build, it's sent to Sentry, so we should be able to
   find it there.

 * It's also sent to `console.log`, so it can be pulled off with either
   `adb logcat` for Android, or the appropriate Xcode feature for iOS.

 * If we want to, it wouldn't be hard to build a couple of additional
   features to make this easier: we could put a switch for that flag
   `enableWebViewErrorDisplay` on a settings screen somewhere; and we
   could extend `logErrorRemotely` (which should probably already get
   a more generic name) to store the error record in an in-app buffer
   which is displayed on a diagnostics screen.  Either of these would
   allow the user to send a screenshot with the information, just like
   before.

As it is, these error messages haven't been super helpful for us even
when we do get them; they're pretty cryptic for us too.  So while
those features would be nice, I don't think we need to wait for them
before shipping this change.