notifications: Fold a single-use type into the signature that uses it.
This used to be exported because of the separation between
platform-specific files and a platform-independent file.
Now that constraint no longer applies.

Folding this into the signature means that hover info on call sites of
`initializeNotifications` now shows the actual contents of the type
rather than this alias.  Which is extra useful given that much of the
information in this type is the names of the three parameters, as they
all have the same type `string`.