storage: Destructure correctly, fixing no-sync bug.
When this code was refactored yesterday in 5075ff66f to make room to
add the `payload &&` condition, a bug was introduced that completely
broke it.  The result is we'll always decide `needsInitialFetch` is
false here... so we won't go fetch any updated data, and won't start
the event-polling loop.

The result is basically the same symptoms as #2800, though what's
going on inside is different (the event loop isn't running at all,
vs. running but just getting the same error again and again.)

Fix it by un-swapping the syntax for a destructuring assignment.