flow: Experiment a bit with "Flow Strict".
This is a setting that's supposed to allow us to turn on a stronger
set of checks in a gradual way across the codebase.

I'd initially hoped it might catch issues like the one fixed in
90b807b06, which I'm kind of mystified why Flow didn't flag in the
first place.  No such luck, though.

It also seems to just not work very well; for example, after adding it
to these files, which were extra easy to include because they had no
imports, I tried adding it to src/emoji/emoticons.js whose only import
is `./emoticonMap`, i.e., one of the files in this commit... and got
"Dependencies of a @flow strict module must also be @flow strict!".

Maybe that last sort of issue is fixed in newer versions of Flow,
though, which we'll be on when we upgrade RN.  So, let's keep this
much around as a demo and we can play again with it later.