ZulipAsyncStorage: Handle null result in getItem.
The underlying AsyncStorage.getItem can return `null` when there's
nothing in the store for this key.  Handle that correctly, passing
it through.

We haven't run into this with redux-persist v4, but it happens with
v5 on first boot-up.

Background: redux-persist v5 handles the initial rehydration
differently.  Previously, the initial rehydration was the job of the
internal function adhocRehydrate().  adhocRehydrate() initialized the
state without accessing the underlying permanent storage.  With v5,
that logic got simplified -- like all other rehydrations, the initial
rehydration initializes the store with what's stored in the permanent
storage.  This means that on the first boot-up, when the permanent
storage is empty, AsyncStorage.getItem() will return `null`.