persist: Refactor slightly to look more like v5-using code.
This is a pure refactor that just reorganizes the code slightly.  This
version is perhaps cleaner in itself; but what really motivates the
change is to make it look more like what we'll have for using
redux-persist v5, to shrink the upcoming diff that does that.

Based on @roberthoenig's drafts in #2762 of the redux-persist upgrade.