persist: Check redux-persist-migrate into the codebase.
redux-persist-migrate allows us to change our state structure while
staying compatible with older app versions.

We check the redux-persist-migrate source directly into our codebase
because it's only 60 LOC and thus easily maintainable.  Furthermore,
upstream development of redux-persist-migrate now happens in
redux-persist v5, which we decided not to use.  To see why we don't
use v5, see: https://github.com/zulip/zulip-mobile/pull/2762

Note that this commit doesn't pass CI on its own, because it is not
compatible with ESLint -- the code is taken as is from
https://github.com/wildlifela/redux-persist-migrate/blob/v5.0.0/src/index.js
The following commit will make the code ESLint compatible. We do this
in a separate commit so that the modifications to the original library
are easily diffable.

The license identification in THIRDPARTY is taken from the author's
package.json file.

[greg: Added THIRDPARTY stanza.  Also created that file just now, in a
 previous commit, to have a place to put that. :-) ]