emoji: Fix misleading variable name `list` for non-lists.
The `Array#reduce` method basically operates on a list... but
these variables are referring to essentially the opposite thing!
The array (or "list") acts as input, but this variable / this
parameter to the inner function is the *output* we're in the
middle of building.

That makes this misnaming super confusing if you don't have the
specifics of the `Array#reduce` signature solid in your head;
and in any event, it's just wrong.  Fix it.