docs/build-run: Document puzzling `libz.so` error and how to fix.
The recipe for the fix comes from
  https://wiki.debian.org/Multiarch/HOWTO#Installing_Android_SDK_compat_libraries
which worked for me on Debian "buster" when setting up my new desktop
a few weeks ago.  Another recipe found on the web and which worked for
another contributor is just the `install` and without the `libgcc1`
package; not sure what variables affect whether that suffices, but
it's just as easy to include it.

Fun fact: on macOS, for comparison, the upcoming 10.14 Mojave
(expected in fall 2018) will be the last version on which 32-bit
libraries will be available at all.