emoji: Improve new selectors the same ways as the old ones.
This means:
* fix misleading, wrong, and merely vague parameter names;
* add types to selectors.

We don't use `RealmEmojiState` for the type because it specifically
calls the keys `id`, and the point of these selectors is that the keys
are not the emojis' IDs but rather their names.  On the other hand,
now that the selector's name explicitly calls out that the result is
indexed by name, it feels redundant to stick that in the type too.