emoji: Start a list of overrides for bad glyphs.
A user reported that :check_mark: wasn't rendering in the app.  Turns
out it renders fine, and shows up clearly in dark mode -- the glyph is
just so light-colored and so lacking in contrast that on a light
background it's effectively invisible!  Fortunately the :check: glyph
is functionally the same, and shows up clearly.