app_refresh: Remove state resetting from several reducers.
At the moment of writing this code we were trying to be extra careful.
This was not an optimal choice in several cases, giving the state
after returning to the app a too strong 'loading' feeling as we are
resetting too much data.

We are doing a full "initial fetch", so this data is going to be
replaced; we don't need to do that early.

* mute - will be refreshed from the `initial fetch`; no need to
  reset at all, the values likely didn't change

* realm - again, will be updated from `initial fetch` and values
  are likely the same

* presence - these likely changed, will be updated from `initial
  fetch`; still makes no sense to throw out the data, as not having
  the data is worse than having it outdated