drafts: Factor "add or remove draft" logic out of ComposeBox.
Add the code from the `tryUpdateDraft` function in ComposeBox as a
separate redux action.  We don't yet actually replace the use of
that function with this one; that'll be a subsequent commit.

Does not include the `if (draft.trim() === message.trim())` line
intentionally, as this is not a behavior we want (and may even
be unintentional reading at the commit that introduced the code).