message: Acknowledge Outbox type in various parameters.
In reality we do pass Outbox values into these functions, as the next
commit will demonstrate, and the code handles them fine; so adjust the
types to acknowledge that.

Prior to aaadc7a32 "outbox: Write an accurate Outbox type", we claimed
that Outbox was a subtype of Message; but that turned out not to be
true as a description of the values we actually pass around.