ui: Rework styling for `FloatingActionButton`
Our `View` component defines the circular shape of the main component.
This makes it not ideal to apply `padding` and `margin` as not only
they might distort the visuals, but they will not expand the touchable
area.

The commit switches the places of `Touchable` and `View` and makes it
so the external style passed is given to the now outer `Touchable`
component.

Functionally this changes nothing. Visually the only difference is
that the touchable effect now spreads out of the button (previously
it was inside but almost invisible).