message: Eliminate NULL_MESSAGE.
A "null object" for a complex application type like this basically
amounts to picking a lot of defaults, many of which can only really
make sense in the context of where they're used.

For example: if there are no messages in a narrow, what should the
"last topic" in the narrow be?  There are a couple of reasonable
values we could have picked; the place where it makes sense to answer
the question, and conversely where a reader is going to have the
question and want to find the answer, is in `getLastTopicForNarrow`.
This moves the answer there.

In the case of `getLastMessageForNarrow`, the decision belongs still
further up the chain -- whatever code wants to use that message object
should have to make a decision about how to handle the case where
there is no message there, so it's best to make that explicit.

(As it happens, that latter function has no callers, so changing the
interface is easy... and also we should just cut it, but I'm doing
that separately so this commit can illustrate the most direct version
of this change.)