Upgrade to React Native v0.57.1.
Holding back from v0.57.2 because of issues there.

Skipping v0.56 because `jest` fails with a Babel config issue (with
or without the Babel config change in this commit), and at this point
v0.56 is old enough it's not worth trying to sort that out.

Both upgrade tools (react-native-git-upgrade and `react-native
upgrade`) failed to run, so instead this was a manual but thorough
old-school upgrade.  Not a problem, since we did that already for all
but the last upgrade.

Most of the changes required for the upgrade have been merged in a
variety of commits over the past few weeks; see #2789.  This commit
 * upgrades `react-native` and closely related dependencies;
 * updates the Flow config file to match RN upstream's template; and
 * adjusts the Babel config to work now that most things use Babel 7:
   move the preset back to our specific babelrc, and re-enable babelrcs.

Fixes #2788.