streamAutocomplete: Eliminate new function creation in render.
StreamItem calls `onPress` with the stream name as parameter, so fit
that paramter in the common function which can be used for all stream
item autocomplete. Before, for every item a new function was created
at the time of rendering.

Create a function inside render should be avoided.
(https://codeburst.io/6-simple-ways-to-speed-up-your-react-native-app-d5b775ab3f16)

This is a expriemental change, if it goes goods then we should do this
at other similiar places too.