title: Mark `narrow` param optional in `getIsInTopicOrStreamNarrow`.
Because `isStreamNarrow` and `isTopicNarrow` handle this case and
return false, this function does the same.

In fact, we already call it with `narrow` undefined: it's called by
`getTitleBackgroundColor` and `getTitleTextColor`, which can be called
from any screen, so if it is not chat screen then narrow will be
undefined.  So fix the type signature to reflect that.