docs/windows: Explain more about why different pieces go where.
Also make a more explicit section for the pieces we're not as
confident of the right recommendation for (like Gradle); and move
Git there, given reports that the filesystem perf limitations hit
it hard.

Also the FAQ, which I don't recall running across when reading
up on this in July, turns out to be pretty great; link to it.