android: Bump Gradle, and Android plugin, to match RN 0.57.
Both new versions represent a major version jump.  Each is
incompatible with our previous version of the other.  There must be
some sequence by which we could upgrade one at a time by taking
smaller steps; but since the upgrade is so simple as is, it's not
worth doing the work to find that.

One very nice feature of plugin version 3.1.4: when dependencies
mention a too-old version of the build tools, it just ignores them!
Version 2.3.3 didn't yet do that, but *did* have a higher minimum
version... which many of our dependencies hadn't updated to.