android: Stop referring to deprecated NDK build system.
The template saw this update in upstream commit 6117a6c72, which
otherwise updated the actual build of RN itself to use NDK r17b.

This line apparently exists to enable a long-deprecated build system:
  https://developer.android.com/studio/projects/add-native-code#ndkCompile
which generates `Android.mk` files as intermediate build products...
of which there are already zero in my build trees, after building the
app.  (And in RN itself, there are plenty of `Android.mk` files in
the source tree instead.)  So it already doesn't seem to be actually
doing anything.

Also since this makes us aware that upstream finally bumped their
NDK version, make a note of that in the one place where we mention
the old one.