compose box: Unify two platforms' code.
Closes #3017.  Fixes #3053.

Remove the iOS code which was our old 'controlled' version and
keep the new Android 'uncontrolled' version.

While iOS does not have the extreme performance issues the Android
app had because of bugs in React Native, it is still faster to not
have the component controlled - a controlled component would block
any time the UI thread blocks.

Another big benefit is removing the need to support two versions.