deps: Upgrade Babel 7 to latest, 7.1.2.
We already have Babel 7 (as `@babel/core` and friends) in addition to
Babel 6 (as `babel-core` and friends), because the `metro@0.30.2`
inside `react-native` requires *both*.  Other reverse-dependencies of
Babel all refer to Babel 6 only.

But the Babel 7 we've been using is a mix of several beta versions.
Upgrade onto fully-released versions, in hopes of smoothing the
upgrade path for moving our Babel 6 usages to Babel 7.  This also
clears a couple of "incorrect peer dependency" warnings we've been
getting every time we run `yarn`.

Doing this completely and cleanly was a bit annoying:

 * I added @babel/core@^7.1.2 (the latest) to package.json,
   for the sake of explicitness.

 * Then running `yarn` upgraded @babel/core and some closely
   related packages, but not other Babel-related packages.
   I wanted to upgrade those too, so:

   * I tried `yarn upgrade --scope @babel`; which *should* do
     the job, AFAICT from the docs, but didn't do much.  I suspect
     it only operates on packages directly named in our package.json.

   * So I did a plain `yarn upgrade`, which upgrades *everything*.
     Then I used `git checkout -p` to take only the changes under
     @babel.

   * Then I ran `yarn`, to pick up the handful of Babel's dependencies
     outside the `@babel` scope which needed upgrades too.

 * I ran `yarn update-libdefs` as usual, which created a new
   stub libdef in flow-typed/.