android notif: Set a basic single channel, to unbreak Oreo (fix #3076)
This basically just copies the sample code from the example in
upstream's guide:
  https://developer.android.com/training/notify-user/build-notification
with small modifications.

There's much more we could and should do to take full advantage of
this spiffy feature; see #2691.  For now, this is a quick box-ticking
fix to get us back to the level of functionality we had before bumping
our `targetSdkVersion` to 26.