webview: Use a round-trip handshake to see when postMessage is ready.
In the message list, we use `postMessage` to communicate between the
JS code running inside the message list's WebView and the JS code
running in the React Native environment.  In the moments after the
WebView is first created, the `postMessage` channel may not yet be
functioning; so we watch for an indication that it is, and queue up
early messages to be sent then.  See b06d2cdd9 and its parent for
where we introduced this.

The indication we watch for is that `window.postMessage` inside the
WebView has a `length` property of 1 (meaning it takes 1 required
argument), as a sign that RN has injected its modified `postMessage`
function.  This worked because the browser's own `window.postMessage`
had a length of 2.

Starting with Chrome 70, the browser's `window.postMessage` has a
length of 1, just like RN's modified version.  So our test succeeds
immediately, and we prematurely signal that `postMessage` is ready.
The app never gets the signal, and never sends messages into the
WebView.  This is #3080, with severe symptoms like #3078.

Fix the problem by switching to a round-trip, end-to-end handshake:
keep sending "ready" messages until we hear one echoed back.

Fixes #3080.
Fixes #3078.

[greg: diagnosed `window.postMessage.length` change;
 wrote commit message.]